Improvements for React hooks
Controllers are now memoized so they don’t trigger re-renders unless the value changes. They also always use the latest value, and don’t have to wait for a re-render to use the latest value. This last part is a significant change in behaviour, but one that should make using the controller much more consistent with hooks… if you use the same controller twice before a re-render you’ll not clobber the previous update anymore.