Add combo snapshot and controller functions so they can be called using values of union types
When you call one of the functions, the arguments have to match at least ONE of the call signatures. Before adding the combo call signature, if you had a value that was a union type like “could be `this` or could be a property name” then you couldn’t call snapshot.